feat: export food with new code
