+++ ecobalyse_data/export/process.py
@@ -1,8 +1,6 @@
-import json
 import os
 from typing import List
 
-import orjson
 from frozendict import frozendict
 
 from common import (
@@ -19,61 +17,7 @@ from common.impacts import impacts as impacts_py
 from common.impacts import main_method
 from ecobalyse_data.computation import compute_impacts, compute_processes_for_activities
 from ecobalyse_data.logging import logger
-from models.process import ComputedBy, Material, Process
-
-
-def activities_to_materials_json(
-    activities_path: str, materials_paths: List[str]
-) -> List[Material]:
-    logger.info(f"-> Loading activities file {activities_path}")
-
-    activities = []
-    with open(activities_path, "r") as file:
-        activities = json.load(file)
-
-    materials = activities_to_materials(activities)
-
-    materials_dict = [material.model_dump() for material in materials]
-
-    exported_files = []
-    for materials_path in materials_paths:
-        with open(materials_path, "wb") as output_file:
-            output_file.write(
-                orjson.dumps(
-                    materials_dict, option=orjson.OPT_INDENT_2 | orjson.OPT_SORT_KEYS
-                )
-            )
-
-            exported_files.append(materials_path)
-
-    format_json(" ".join(exported_files))
-
-    for materials_path in exported_files:
-        logger.info(f"-> Exported {len(materials_dict)} materials to {materials_path}")
-
-
-def activities_to_materials(activities: List[dict]) -> List[Material]:
-    return [
-        activity_to_material(activity)
-        for activity in list(activities)
-        if activity["category"] == "material"
-    ]
-
-
-def activity_to_material(activity: dict) -> Material:
-    return Material(
-        id=activity["material_id"],
-        materialProcessUuid=activity["id"],
-        recycledProcessUuid=activity.get("recycledProcessUuid"),
-        recycledFrom=activity.get("recycledFrom"),
-        name=activity["shortName"],
-        shortName=activity["shortName"],
-        origin=activity["origin"],
-        primary=activity.get("primary"),
-        geographicOrigin=activity["geographicOrigin"],
-        defaultCountry=activity["defaultCountry"],
-        cff=activity.get("cff"),
-    )
+from models.process import ComputedBy, Process
 
 
 def get_activities_by_id(activities, logger) -> frozendict:
@@ -82,7 +26,7 @@ def get_activities_by_id(activities, logger) -> frozendict:
 
 
 def activities_to_processes(
-    activities_path: str,
+    activities: list[dict],
     aggregated_relative_file_path: str,
     impacts_relative_file_path: str,
     dirs_to_export_to: List[str],
@@ -91,12 +35,6 @@ def activities_to_processes(
     verbose: bool = False,
     display_changes: bool = True,
 ):
-    logger.debug(f"-> Loading activities file {activities_path}")
-
-    activities = []
-    with open(activities_path, "r") as file:
-        activities = json.load(file)
-
     factors = get_normalization_weighting_factors(IMPACTS_JSON)
     processes: List[Process] = compute_processes_for_activities(
         activities, main_method, impacts_py, IMPACTS_JSON, factors
